+++ lib/capture_settings.dart
@@ -1,5 +1,0 @@
-class CaptureSettings {
-  Map<String, dynamic> toMap() {
-    return {};
-  }
-}