+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.10.0
+* Added sessionPreset to change camera quality in CaptureSettings for ios only.
+* Bump up HaishinKit.kt to 0.10.0
+* Feature: add setCaptureSettings
+
+## 0.9.4
+* Closing & disposing fixes. Fixed android camera facing.
+* Fix: Android - Pass audioSettings.muted to RtmpStream.
+
 ## 0.9.3
 * Bump up HaishinKit.kt to 0.10.4
 * Improved compatibility some services.